requirements.txt
tensorflow
numpy
scipy
streamlit
pillow
python-dotenv
requests
tqdm
kagglehub
